Создает модуль функций
